Module libffi::high::arity0

CIF and closure types organized by function arity.

Structs

Cif0

A typed CIF, which statically tracks argument and result types.

Closure0

An immutable, typed closure with the given argument and result types.

ClosureMut0

A mutable, typed closure with the given argument and result types.

ClosureOnce0

A one-shot, typed closure with the given argument and result types.

FnPtr0

A lifetime carrying wrapper type for fn pointers.

Type Aliases

Callback0

The type of function called from an immutable, typed closure.

CallbackMut0

The type of function called from a mutable, typed closure.

CallbackOnce0

The type of function called from a one-shot, typed closure.