Module libffi::high::arity10

CIF and closure types organized by function arity.

Structs

Cif10
A typed CIF, which statically tracks argument and result types.

Closure10
An immutable, typed closure with the given argument and result types.

ClosureMut10
A mutable, typed closure with the given argument and result types.

ClosureOnce10
A one-shot, typed closure with the given argument and result types.

Type Definitions

Callback10
The type of function called from an immutable, typed closure.

CallbackMut10
The type of function called from a mutable, typed closure.

CallbackOnce10
The type of function called from a one-shot, typed closure.