Struct libffi::high::arity11::ClosureOnce11

pub struct ClosureOnce11<A, B, C, D, E, F, G, H, I, J, K, R> { /* fields omitted */ }

A one-shot, typed closure with the given argument and result types.

Methods

impl<A: CType, B: CType, C: CType, D: CType, E: CType, F: CType, G: CType, H: CType, I: CType, J: CType, K: CType, R: CType> ClosureOnce11<A, B, C, D, E, F, G, H, I, J, K, R>

pub fn new<Callback>(callback: Callback) -> Self where
Callback: FnOnce(A, B, C, D, E, F, G, H, I, J, K) -> R + Any,

Constructs a typed closure callable from C from a Rust closure.

impl<A: Copy, B: Copy, C: Copy, D: Copy, E: Copy, F: Copy, G: Copy, H: Copy, I: Copy, J: Copy, K: Copy, R> ClosureOnce11<A, B, C, D, E, F, G, H, I, J, K, R>

pub fn new_with_cif<Callback>(
cif: Cif11<A, B, C, D, E, F, G, H, I, J, K, R>,
callback: Callback
) -> Self where
Callback: FnOnce(A, B, C, D, E, F, G, H, I, J, K) -> R + Any,

Constructs a one-shot closure callable from C from a CIF describing the calling convention for the resulting function and the Rust closure to call.

impl<A, B, C, D, E, F, G, H, I, J, K, R> ClosureOnce11<A, B, C, D, E, F, G, H, I, J, K, R>

pub fn code_ptr(
&self
) -> &extern "C" fn(_: A, _: B, _: C, _: D, _: E, _: F, _: G, _: H, _: I, _: J, _: K) -> R

Gets the C code pointer that is used to invoke the closure.

pub fn from_parts<U: Any>(
cif: Cif11<A, B, C, D, E, F, G, H, I, J, K, R>,
callback: CallbackOnce11<U, A, B, C, D, E, F, G, H, I, J, K, R>,
userdata: U
) -> Self

Constructs a one-shot closure callable from C from a CIF describing the calling convention for the resulting function, a callback for the function to call, and userdata to pass to the callback.