Struct libffi::high::arity12::Cif12

pub struct Cif12<A, B, C, D, E, F, G, H, I, J, K, L, R> { /* fields omitted */ }

A typed CIF, which statically tracks argument and result types.

Implementations

impl<A, B, C, D, E, F, G, H, I, J, K, L, R> Cif12<A, B, C, D, E, F, G, H, I, J, K, L, R>

pub fn new(
A: Type<A>,
B: Type<B>,
C: Type<C>,
D: Type<D>,
E: Type<E>,
F: Type<F>,
G: Type<G>,
H: Type<H>,
I: Type<I>,
J: Type<J>,
K: Type<K>,
L: Type<L>,
result: Type<R>
) -> Self

Creates a new statically-typed CIF with the given argument and result types.

pub fn set_abi(&mut self, abi: FfiAbi)

Sets the CIF to use the given calling convention.

impl<A: CType, B: CType, C: CType, D: CType, E: CType, F: CType, G: CType, H: CType, I: CType, J: CType, K: CType, L: CType, R: CType> Cif12<A, B, C, D, E, F, G, H, I, J, K, L, R>

pub fn reify() -> Self

Creates a new statically-typed CIF by reifying the argument types as Type<T>s.