Module libffi::high::arity2

CIF and closure types organized by function arity.

Structs

Cif2
A typed CIF, which statically tracks argument and result types.

Closure2
An immutable, typed closure with the given argument and result types.

ClosureMut2
A mutable, typed closure with the given argument and result types.

ClosureOnce2
A one-shot, typed closure with the given argument and result types.

Type Definitions

Callback2
The type of function called from an immutable, typed closure.

CallbackMut2
The type of function called from a mutable, typed closure.

CallbackOnce2
The type of function called from a one-shot, typed closure.