Struct libffi::high::arity5::Closure5

pub struct Closure5<'a, A, B, C, D, E, R> { /* fields omitted */ }

An immutable, typed closure with the given argument and result types.

Methods

impl<'a, A: CType, B: CType, C: CType, D: CType, E: CType, R: CType> Closure5<'a, A, B, C, D, E, R>

pub fn new<Callback>(callback: &'a Callback) -> Self where
Callback: Fn(A, B, C, D, E) -> R + 'a,

Constructs a typed closure callable from C from a Rust closure.

impl<'a, A, B, C, D, E, R> Closure5<'a, A, B, C, D, E, R>

pub fn code_ptr(&self) -> &extern "C" fn(_: A, _: B, _: C, _: D, _: E) -> R

Gets the C code pointer that is used to invoke the closure.

pub fn from_parts<U>(
cif: Cif5<A, B, C, D, E, R>,
callback: Callback5<U, A, B, C, D, E, R>,
userdata: &'a U
) -> Self

Constructs a typed closure callable from C from a CIF describing the calling convention for the resulting function, a callback for the function to call, and userdata to pass to the callback.

impl<'a, A: Copy, B: Copy, C: Copy, D: Copy, E: Copy, R> Closure5<'a, A, B, C, D, E, R>

pub fn new_with_cif<Callback>(
cif: Cif5<A, B, C, D, E, R>,
callback: &'a Callback
) -> Self where
Callback: Fn(A, B, C, D, E) -> R + 'a,

Constructs a typed closure callable from C from a CIF describing the calling convention for the resulting function and the Rust closure to call.