Struct libffi::high::arity7::Closure7

pub struct Closure7<'a, A, B, C, D, E, F, G, R> {
    // some fields omitted
}

An immutable, typed closure with the given argument and result types.

Methods

impl<'a, A: CType, B: CType, C: CType, D: CType, E: CType, F: CType, G: CType, R: CType> Closure7<'a, A, B, C, D, E, F, G, R>

fn new<Callback>(callback: &'a Callback) -> Self where Callback: Fn(A, B, C, D, E, F, G) -> R + 'a

Constructs a typed closure callable from C from a Rust closure.

impl<'a, A, B, C, D, E, F, G, R> Closure7<'a, A, B, C, D, E, F, G, R>

fn code_ptr(&self) -> & extern "C" fn(A, B, C, D, E, F, G) -> R

Gets the C code pointer that is used to invoke the closure.

fn from_parts<U>(cif: Cif7<A, B, C, D, E, F, G, R>, callback: Callback7<U, A, B, C, D, E, F, G, R>, userdata: &'a U) -> Self

Constructs a typed closure callable from C from a CIF describing the calling convention for the resulting function, a callback for the function to call, and userdata to pass to the callback.

impl<'a, A: Copy, B: Copy, C: Copy, D: Copy, E: Copy, F: Copy, G: Copy, R> Closure7<'a, A, B, C, D, E, F, G, R>

fn new_with_cif<Callback>(cif: Cif7<A, B, C, D, E, F, G, R>, callback: &'a Callback) -> Self where Callback: Fn(A, B, C, D, E, F, G) -> R + 'a

Constructs a typed closure callable from C from a CIF describing the calling convention for the resulting function and the Rust closure to call.