Module libffi::high::arity8

CIF and closure types organized by function arity.

Structs

Cif8

A typed CIF, which statically tracks argument and result types.

Closure8

An immutable, typed closure with the given argument and result types.

ClosureMut8

A mutable, typed closure with the given argument and result types.

ClosureOnce8

A one-shot, typed closure with the given argument and result types.

FnPtr8

A lifetime carrying wrapper type for fn pointers.

Type Aliases

Callback8

The type of function called from an immutable, typed closure.

CallbackMut8

The type of function called from a mutable, typed closure.

CallbackOnce8

The type of function called from a one-shot, typed closure.