libffi::high::arity9::Callback9

type Callback9<U, A, B, C, D, E, F, G, H, I, R> =  extern "C" fn(cif: &ffi_cif, result: &mut R, args: &(&A, &B, &C, &D, &E, &F, &G, &H, &I), userdata: &U);

The type of function called from an immutable, typed closure.