Struct libffi::high::call::Arg

pub struct Arg<'a> { /* fields omitted */ }

Encapsulates an argument with its type information.

In order to set up calls using call, we need to wrap (a reference to) each argument in an Arg. The usual way to do this is with function arg.

Methods

impl<'a> Arg<'a>

pub fn new<T: CType>(arg: &'a T) -> Self

Wraps an argument reference for passing to high::call::call.

For a shorter alias of the same, see high::call::arg.