Module libffi::low::types

Re-exports the ffi_type objects used to describe the types of arguments and results.

These are from raw, but are renamed by removing the ffi_type_ prefix. For example, raw::ffi_type_void becomes low::types::void.

Reexports

pub use raw::{ffi_type_void as void, ffi_type_uint8 as uint8, ffi_type_sint8 as sint8, ffi_type_uint16 as uint16, ffi_type_sint16 as sint16, ffi_type_uint32 as uint32, ffi_type_sint32 as sint32, ffi_type_uint64 as uint64, ffi_type_sint64 as sint64, ffi_type_float as float, ffi_type_double as double, ffi_type_pointer as pointer, ffi_type_longdouble as longdouble, ffi_type_complex_float as complex_float, ffi_type_complex_double as complex_double, ffi_type_complex_longdouble as complex_longdouble};