Struct libffi::middle::ClosureOnce

pub struct ClosureOnce {
    // some fields omitted
}

A closure that owns needs-drop data.

This allows the closure’s callback to take ownership of the data, in which case the userdata will be gone if called again.

Methods

impl ClosureOnce

fn new<U: Any, R>(cif: Cif, callback: CallbackOnce<U, R>, userdata: U) -> Self

Creates a new closure with owned userdata.

Arguments

cif — describes the calling convention and argument and result types

callback — the function to call when the closure is invoked

userdata — the value to pass to callback along with the arguments when the closure is called

Result

The new closure.

fn code_ptr(&self) -> &unsafe extern "C" fn()

Obtains the callable code pointer for a closure.

Safety

The result needs to be transmuted to the correct type before it can be called. If the type is wrong then undefined behavior will result.