Type Definition libffi::high::arity8::Callback8

type Callback8<U, A, B, C, D, E, F, G, H, R> = extern "C" fn(cif: &ffi_cif, result: &mut R, args: &(&A, &B, &C, &D, &E, &F, &G, &H), userdata: &U);

The type of function called from an immutable, typed closure.