Trait libffi::high::types::CType

pub unsafe trait CType: Copy {
    fn reify() -> Type<Self>;
}

Types that we can automatically marshall to/from C.

In particular, for any type T that implements CType, we can get a Type<T> for describing that type. This trait is unsafe to implement because if the libffi type associated with a Rust type doesn’t match then we get undefined behavior.

Required Methods

fn reify() -> Type<Self>

Creates or retrieves a Type<T> for any type T: CType.

We can use the resulting object to assemble a CIF to set up a call that uses type T.