Module libffi::low::types

Re-exports the ffi_type objects used to describe the types of arguments and results.

These are from the raw layer, but are renamed by removing the ffi_type_ prefix. For example, raw::ffi_type_void becomes types::void.

Statics

double

float

pointer

sint8

sint16

sint32

sint64

uint8

uint16

uint32

uint64

void